test-fridge1/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
ultralytics>=8.0.0
opencv-python>=4.6.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
Pillow>=9.0.0

